Release/4.7.0 (#128)
* Emit wallet balance information to New Relic (#126)

* Add balanceOf method in Blockchain

Add a method in the Blockchain class responsible for fetching the balance of a wallet given the address.

* Add Homechain Balance Reporter

* Add Foreign Chain Balance Reporter

* Enable balance reporters in MetricsWorker

* Update event recorders

* Add BalanceReporter

Create an abstract class BalanceReporter to avoid duplicating code

* Add ethereum chain name formatter

Due to standardization, change 'ethereum' to 'eth' when logging and notifying

* Apply code review suggestions

* Remove class state mutation

This actually should be done in the constructor, but it wasn't working with inversify so I made these methods. Big sorry.

* Update bigNumberToBalance formatter

* Refactor Balance Reporters

Turned arrow functions into traditional functions where possible. Add postConstruct setup to HomechainBalanceReporter

* Update bigNumberToBalance

* Refactor chains currencies to enum

* Refactor bigNumberToBalance

* Code review suggestions

* Decrease balance decimal cases to two

* Add Avalanche to balance reporter

Co-authored-by: Pedro Ramos <pedro.henrique@codeminer42.com>

* Update version

* Use ForeignChainsIds object

Avoid code duplication by importing the ForeignChainsIds object

* Update CHANGELOG.md

* Add validator URL prop

* Implement code review suggestions

Co-authored-by: marcelow <marcelow@internet>
Co-authored-by: DZariusz <34737849+DZariusz@users.noreply.github.com>